Move Stylecop checks to their own helper executable.
