+++ make.ps1
@@ -137,24 +137,32 @@ function Check-Command {
 	{
 		echo "Checking for explicit interface violations..."
 		./OpenRA.Utility.exe all --check-explicit-interfaces
-		echo "Checking for code style violations in OpenRA.Platforms.Default..."
-		./OpenRA.Utility.exe cnc --check-code-style OpenRA.Platforms.Default
-		echo "Checking for code style violations in OpenRA.Game..."
-		./OpenRA.Utility.exe ra --check-code-style OpenRA.Game
-		echo "Checking for code style violations in OpenRA.Mods.Common..."
-		./OpenRA.Utility.exe ra --check-code-style OpenRA.Mods.Common
-		echo "Checking for code style violations in OpenRA.Mods.Cnc..."
-		./OpenRA.Utility.exe cnc --check-code-style OpenRA.Mods.Cnc
-		echo "Checking for code style violations in OpenRA.Mods.D2k..."
-		./OpenRA.Utility.exe cnc --check-code-style OpenRA.Mods.D2k
-		echo "Checking for code style violations in OpenRA.Utility..."
-		./OpenRA.Utility.exe cnc --check-code-style OpenRA.Utility
-		echo "Checking for code style violations in OpenRA.Test..."
-		./OpenRA.Utility.exe cnc --check-code-style OpenRA.Test
 	}
 	else
 	{
 		UtilityNotFound
+	}
+
+	if (Test-Path OpenRA.StyleCheck.exe)
+	{
+		echo "Checking for code style violations in OpenRA.Platforms.Default..."
+		./OpenRA.StyleCheck.exe OpenRA.Platforms.Default
+		echo "Checking for code style violations in OpenRA.Game..."
+		./OpenRA.StyleCheck.exe OpenRA.Game
+		echo "Checking for code style violations in OpenRA.Mods.Common..."
+		./OpenRA.StyleCheck.exe OpenRA.Mods.Common
+		echo "Checking for code style violations in OpenRA.Mods.Cnc..."
+		./OpenRA.StyleCheck.exe OpenRA.Mods.Cnc
+		echo "Checking for code style violations in OpenRA.Mods.D2k..."
+		./OpenRA.StyleCheck.exe OpenRA.Mods.D2k
+		echo "Checking for code style violations in OpenRA.Utility..."
+		./OpenRA.StyleCheck.exe OpenRA.Utility
+		echo "Checking for code style violations in OpenRA.Test..."
+		./OpenRA.StyleCheck.exe OpenRA.Test
+	}
+	else
+	{
+		echo "OpenRA.StyleCheck.exe could not be found. Build the project first using the `"all`" command."
 	}
 }
